+++ packages/mixin-extension-base/src/background/handlers/extension/keyring.ts
@@ -14,7 +14,7 @@ import { createSubscription, unsubscribe } from "../subscriptions";
 export default function (state: State) {
   return {
     createKeyringStateSubscribe(id: string, port: chrome.runtime.Port) {
-      const cb = createSubscription<"pri(keyring.subscribe)">(id, port);
+      const cb = createSubscription<"pri_(keyring.subscribe)">(id, port);
       const subscription = state.keyring.keyringMemStateSubject.subscribe(
         (state) => {
           cb(state);