+++ packages/mixin-extension-base/src/background/types/wallet.ts
@@ -28,14 +28,14 @@ export interface MultiSigsPaymentPayload {
 }
 
 export interface WalletActionSignatures {
-  "pri(transfer.list)": [null, boolean, TransferReq[]];
-  "pri(transfer.approve)": [ApproveTransferPayload, boolean];
-  "pri(transfer.reject)": [RejectTransferPayload, boolean];
-  "pri(multisigs.list)": [null, boolean, RawTransactionReq[]];
-  "pri(multisigs.approve)": [ApproveMultisigsPayload, boolean];
-  "pri(multisigs.reject)": [RejectMultisigsPayload, boolean];
+  "pri_(transfer.list)": [null, boolean, TransferReq[]];
+  "pri_(transfer.approve)": [ApproveTransferPayload, boolean];
+  "pri_(transfer.reject)": [RejectTransferPayload, boolean];
+  "pri_(multisigs.list)": [null, boolean, RawTransactionReq[]];
+  "pri_(multisigs.approve)": [ApproveMultisigsPayload, boolean];
+  "pri_(multisigs.reject)": [RejectMultisigsPayload, boolean];
 
-  "pub(transfer.request)": [CreateTransferPayload, null];
-  "pub(multisigs.request)": [MultiSigsPaymentPayload, boolean];
-  "pub(token.sign)": [SignClientTokenPayload, string];
+  "pub_(transfer.request)": [CreateTransferPayload, null];
+  "pub_(multisigs.request)": [MultiSigsPaymentPayload, boolean];
+  "pub_(token.sign)": [SignClientTokenPayload, string];
 }