+++ packages/mixin-extension-ui/src/messages/platform.ts
@@ -3,11 +3,11 @@ import { SendMessage } from "@foxone/mixin-extension-base/inject/types";
 export default function (sendMessage: SendMessage) {
   return {
     closePopup() {
-      return sendMessage("pri(platform.closePopup)");
+      return sendMessage("pri_(platform.closePopup)");
     },
 
     openWindow(path: string) {
-      return sendMessage("pri(platform.openWindow)", { path });
+      return sendMessage("pri_(platform.openWindow)", { path });
     }
   };
 }