+++ packages/mixin-extension-ui/src/messages/wallet.ts
@@ -7,31 +7,31 @@ import {
 export default function (sendMessage: SendMessage) {
   return {
     rejectTransferReq(id: string): Promise<boolean> {
-      return sendMessage("pri(transfer.reject)", { id });
+      return sendMessage("pri_(transfer.reject)", { id });
     },
 
     approveTransferReq(id: string): Promise<boolean> {
-      return sendMessage("pri(transfer.approve)", { id });
+      return sendMessage("pri_(transfer.approve)", { id });
     },
 
     subscribeTransferReq(
       cb: (requests: TransferReq[]) => void
     ): Promise<boolean> {
-      return sendMessage("pri(transfer.list)", null, cb);
+      return sendMessage("pri_(transfer.list)", null, cb);
     },
 
     rejectMultisigsTransactionReq(id: string): Promise<boolean> {
-      return sendMessage("pri(multisigs.reject)", { id });
+      return sendMessage("pri_(multisigs.reject)", { id });
     },
 
     approveMultisigsTransactionReq(id: string): Promise<boolean> {
-      return sendMessage("pri(multisigs.approve)", { id });
+      return sendMessage("pri_(multisigs.approve)", { id });
     },
 
     subscribeMultisigsTransactionReq(
       cb: (requests: RawTransactionReq[]) => void
     ): Promise<boolean> {
-      return sendMessage("pri(multisigs.list)", null, cb);
+      return sendMessage("pri_(multisigs.list)", null, cb);
     }
   };
 }