+++ packages/mixin-extension/webpack.base.js
@@ -5,11 +5,10 @@ const CopyPlugin = require("copy-webpack-plugin");
 const VueLoaderPlugin = require("vue-loader/lib/plugin");
 const { CleanWebpackPlugin } = require("clean-webpack-plugin");
 const ProgressBarPlugin = require("progress-bar-webpack-plugin");
-const WextManifestWebpackPlugin = require('wext-manifest-webpack-plugin');
+const ManifestPlugin = require("webpack-extension-manifest-plugin");
 
 const pkgJson = require("../../lerna.json");
-const destPath = path.join(__dirname, 'build');
-const targetBrowser = process.env.TARGET_BROWSER;
+const manifest = require("./manifest.json");
 
 const packages = [
   "mixin-extension",
@@ -21,22 +20,8 @@ const packages = [
 module.exports = {
   context: __dirname,
   devtool: false,
-  // entry: {
-  //     manifest: './manifest.json',
-  // },
   module: {
     rules: [
-      {
-        type: 'javascript/auto', // prevent webpack handling json with its own loaders,
-        test: /manifest\.json$/,
-        use: {
-            loader: 'wext-manifest-loader',
-            options: {
-                usePackageJSONVersion: true,
-            },
-        },
-        exclude: /node_modules/,
-      },
       {
         test: /\.(js|ts|tsx)$/,
         exclude: [/node_modules/],
@@ -105,7 +90,7 @@ module.exports = {
     chunkFilename: "[name].js",
     filename: "[name].js",
     globalObject: "(typeof self !== 'undefined' ? self : this)",
-    path: path.join(destPath, targetBrowser)
+    path: path.join(__dirname, "build")
   },
 
   performance: {
@@ -113,7 +98,6 @@ module.exports = {
   },
 
   plugins: [
-    new WextManifestWebpackPlugin(),
     new webpack.ProvidePlugin({
       Buffer: ["buffer", "Buffer"],
       process: "process/browser.js"
@@ -125,6 +109,15 @@ module.exports = {
       }
     }),
     new CopyPlugin({ patterns: [{ from: "public" }] }),
+    new ManifestPlugin({
+      config: {
+        base: manifest,
+        extend: {
+          version: pkgJson.version.split("-")[0], // remove possible -beta.xx
+          version_name: pkgJson.version
+        }
+      }
+    }),
     new VueLoaderPlugin(),
     new CleanWebpackPlugin(),
     new ProgressBarPlugin()