+++ packages/mixin-extension/webpack.config.js
@@ -1,7 +1,6 @@
 const base = require("./webpack.base");
 
 const entry = {
-  manifest: './manifest.json',
   extension: "./src/extension.ts",
   background: "./src/background.ts",
   content: "./src/content.ts",