+++ packages/mixin-extension-ui/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.12](https://github.com/fox-one/mixin-extension/compare/v0.1.0-alpha.11...v0.1.0-alpha.12) (2021-04-14)
+
+**Note:** Version bump only for package @foxone/mixin-extension-ui
+
+
+
+
+
+# [0.1.0-alpha.11](https://github.com/fox-one/mixin-extension/compare/v0.1.0-alpha.10...v0.1.0-alpha.11) (2021-04-14)
+
+**Note:** Version bump only for package @foxone/mixin-extension-ui
+
+
+
+
+
+# [0.1.0-alpha.10](https://github.com/fox-one/mixin-extension/compare/v0.1.0-alpha.9...v0.1.0-alpha.10) (2021-04-14)
+
+**Note:** Version bump only for package @foxone/mixin-extension-ui
+
+
+
+
+
 # [0.1.0-alpha.9](https://github.com/fox-one/mixin-extension/compare/v0.1.0-alpha.8...v0.1.0-alpha.9) (2021-04-14)